# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto signal detector API powered by CoinGecko
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CoinSignalSignal = typeof CoinSignalSignal[keyof typeof CoinSignalSignal];


export const CoinSignalSignal = {
  buy_long: 'buy_long',
  sell_short: 'sell_short',
  neutral: 'neutral',
} as const;

export interface CoinSignal {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  /** @nullable */
  price_change_percentage_24h: number | null;
  /** @nullable */
  price_change_percentage_7d_in_currency: number | null;
  total_volume: number;
  /** @nullable */
  ath?: number | null;
  /** @nullable */
  ath_change_percentage?: number | null;
  signal: CoinSignalSignal;
  /** Signal strength from -100 (strong sell) to +100 (strong buy) */
  signal_strength: number;
  signal_reasons: string[];
}

export type CoinDetailSignal = typeof CoinDetailSignal[keyof typeof CoinDetailSignal];


export const CoinDetailSignal = {
  buy_long: 'buy_long',
  sell_short: 'sell_short',
  neutral: 'neutral',
} as const;

export interface CoinDetail {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  /** @nullable */
  price_change_percentage_24h: number | null;
  /** @nullable */
  price_change_percentage_7d_in_currency: number | null;
  /** @nullable */
  price_change_percentage_30d_in_currency?: number | null;
  total_volume: number;
  /** @nullable */
  ath?: number | null;
  /** @nullable */
  ath_change_percentage?: number | null;
  /** @nullable */
  circulating_supply?: number | null;
  /** @nullable */
  total_supply?: number | null;
  signal: CoinDetailSignal;
  signal_strength: number;
  signal_reasons: string[];
  sparkline_in_7d: number[];
}

export interface SignalGroups {
  buy_long: CoinSignal[];
  sell_short: CoinSignal[];
  neutral: CoinSignal[];
}

export type MarketSummarySignalCounts = {
  buy_long: number;
  sell_short: number;
  neutral: number;
};

export interface MarketSummary {
  total_market_cap_usd: number;
  total_volume_24h: number;
  btc_dominance: number;
  market_cap_change_24h: number;
  signal_counts: MarketSummarySignalCounts;
}

export interface TopSignals {
  top_buy_long: CoinSignal[];
  top_sell_short: CoinSignal[];
}

export type ListCoinsParams = {
vs_currency?: string;
per_page?: number;
/**
 * Filter by signal type
 * @nullable
 */
signal?: ListCoinsSignal;
};

export type ListCoinsSignal = typeof ListCoinsSignal[keyof typeof ListCoinsSignal] | null;


export const ListCoinsSignal = {
  buy_long: 'buy_long',
  sell_short: 'sell_short',
  neutral: 'neutral',
} as const;