# lib/api-zod/src/generated/types/coinDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto signal detector API powered by CoinGecko
 * OpenAPI spec version: 0.1.0
 */
import type { CoinDetailSignal } from './coinDetailSignal';

export interface CoinDetail {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  /** @nullable */
  price_change_percentage_24h: number | null;
  /** @nullable */
  price_change_percentage_7d_in_currency: number | null;
  /** @nullable */
  price_change_percentage_30d_in_currency?: number | null;
  total_volume: number;
  /** @nullable */
  ath?: number | null;
  /** @nullable */
  ath_change_percentage?: number | null;
  /** @nullable */
  circulating_supply?: number | null;
  /** @nullable */
  total_supply?: number | null;
  signal: CoinDetailSignal;
  signal_strength: number;
  signal_reasons: string[];
  sparkline_in_7d: number[];
}