# lib/api-zod/src/generated/types/coinSignal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto signal detector API powered by CoinGecko
 * OpenAPI spec version: 0.1.0
 */
import type { CoinSignalSignal } from './coinSignalSignal';

export interface CoinSignal {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  /** @nullable */
  price_change_percentage_24h: number | null;
  /** @nullable */
  price_change_percentage_7d_in_currency: number | null;
  total_volume: number;
  /** @nullable */
  ath?: number | null;
  /** @nullable */
  ath_change_percentage?: number | null;
  signal: CoinSignalSignal;
  /** Signal strength from -100 (strong sell) to +100 (strong buy) */
  signal_strength: number;
  signal_reasons: string[];
}