# lib/api-zod/src/generated/types/marketSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto signal detector API powered by CoinGecko
 * OpenAPI spec version: 0.1.0
 */
import type { MarketSummarySignalCounts } from './marketSummarySignalCounts';

export interface MarketSummary {
  total_market_cap_usd: number;
  total_volume_24h: number;
  btc_dominance: number;
  market_cap_change_24h: number;
  signal_counts: MarketSummarySignalCounts;
}