# lib/api-zod/src/generated/types/signalGroups.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto signal detector API powered by CoinGecko
 * OpenAPI spec version: 0.1.0
 */
import type { CoinSignal } from './coinSignal';

export interface SignalGroups {
  buy_long: CoinSignal[];
  sell_short: CoinSignal[];
  neutral: CoinSignal[];
}