# lib/api-zod/src/generated/types/topSignals.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto signal detector API powered by CoinGecko
 * OpenAPI spec version: 0.1.0
 */
import type { CoinSignal } from './coinSignal';

export interface TopSignals {
  top_buy_long: CoinSignal[];
  top_sell_short: CoinSignal[];
}